+++ src/FunFair.CodeAnalysis/Rules.cs
@@ -13,5 +13,6 @@ namespace FunFair.CodeAnalysis
         public const string RuleDontUseAssertTrueWithoutMessage = @"FFS0009";
         public const string RuleDontUseAssertFalseWithoutMessage = @"FFS0010";
         public const string RuleStructsShouldBeReadOnly = @"FFS0011";
+        public const string RuleClassesShouldBeStaticSealedOrAbstract = @"FFS0012";
     }
 }